# contracts/openapi/topsdojob-v3-local.yaml
openapi: 3.1.0
info:
  title: Tops do Job V3 Local API
  version: 0.0.1
  description: "Contrato minimo local do skeleton, camada transversal, API publica de leitura e funil local Anuncie gratis. A API segue fail-closed: endpoints nao descritos em /api/** nao devem ficar publicos por acidente."
servers:
  - url: http://localhost:8080
    description: backend local
paths:
...
  /api/public/minha-conta/anuncios/{slug}:
    get:
      summary: Detalha um anuncio pertencente ao usuario publico autenticado
      operationId: getCurrentUserAd
      security:
        - PublicSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9-]{1,120}$"
      description: O backend resolve o usuario pela sessao e recusa acesso a anuncio pertencente a outro usuario.
      responses:
        "200":
          description: detalhe sanitizado do anuncio do usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeuAnuncio"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Atualiza os campos editaveis de um anuncio do usuario autenticado
      operationId: updateCurrentUserAd
      security:
        - PublicSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/CurrentUserAdSlug"
      description: O proprietario e resolvido exclusivamente pela sessao. Slug, datas de publicacao, proprietario, beneficios e midias nao sao alterados. A edicao remete o anuncio para a revisao vigente e preserva a URL publica.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeuAnuncioAtualizacaoRequest"
      responses:
        "200":
          description: anuncio atualizado e remetido para revisao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeuAnuncio"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      summary: Remove logicamente um anuncio do usuario autenticado
      operationId: removeCurrentUserAd
      security:
        - PublicSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/CurrentUserAdSlug"
      description: O proprietario e resolvido exclusivamente pela sessao. A operacao marca o anuncio como REMOVIDO e preserva dados, midias, historico e objetos de storage.
      responses:
        "200":
          description: anuncio removido logicamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeuAnuncioCicloVida"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/public/minha-conta/anuncios/{slug}/midias:
    get:
      summary: Lista as midias e os limites do anuncio do usuario autenticado
      operationId: listCurrentUserAdMedia
      security:
        - PublicSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/CurrentUserAdSlug"
      responses:
        "200":
          description: Midias do proprio anuncio, sem bucket, object key, hash ou URL publica indevida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinhasMidiasResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Envia uma foto ou video para a area privada do proprio anuncio
      operationId: uploadCurrentUserAdMedia
      security:
        - PublicSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/CurrentUserAdSlug"
        - name: Idempotency-Key
          in: header
          required: false
          description: Obrigatoria para fotos; identifica retries tecnicos do mesmo upload.
          schema:
            type: string
            minLength: 1
            maxLength: 160
            pattern: "^[A-Za-z0-9._:-]+$"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [arquivo]
              properties:
                arquivo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Midia privada criada como pendente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinhasMidiasResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
...
  /api/public/anuncios/{slug}:
    get:
      summary: Detalhe publico minimo de anuncio por slug
      operationId: getPublicAnuncioBySlug
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9-]{1,120}$"
      responses:
        "200":
          description: anuncio publico encontrado
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnuncioDetalhePublico"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/anuncios:
    get:
      summary: Lista administrativa local de anuncios
      description: Retorna pagina sanitizada de anuncios com proprietario mascarado e revisao aberta, somente para ADMIN ou MODERADOR com ANUNCIO_LER. Nao retorna documento privado, e-mail integral, storage ou payload.
      operationId: getAdminAnuncios
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: statusModeracao
          in: query
          required: false
          schema:
            type: string
        - name: uf
          in: query
          required: false
          schema:
            type: string
            pattern: "^[A-Za-z]{2}$"
        - name: cidade
          in: query
          required: false
          schema:
            type: string
        - name: bairro
          in: query
          required: false
          schema:
            type: string
        - name: termo
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
      responses:
        "200":
          description: pagina de anuncios admin read-only
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPaginaAnuncios"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/anuncios/{id}:
    get:
      summary: Detalhe administrativo sanitizado de anuncio
      description: Retorna os dados comerciais necessarios a moderacao, proprietario mascarado e revisao aberta, somente para ADMIN ou MODERADOR com ANUNCIO_LER.
      operationId: getAdminAnuncioDetalhe
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: detalhe sanitizado de anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnuncioDetalhe"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/anuncios/{id}/midias:
    get:
      summary: Lista midias sanitizadas de um anuncio
      description: Restrito a ADMIN e MODERADOR. Retorna somente FOTO e VIDEO; STORY nao participa da fila. Nao retorna bucket, chave, hash, etag, URL privada ou nome original.
      operationId: getAdminAnuncioMidias
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: pagina de midias do anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPaginaMidias"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/anuncios/{id}/historico-moderacao:
    get:
      summary: Historico sanitizado da moderacao de um anuncio
      description: Retorna decisoes do anuncio e de suas fotos e videos. Stories e snapshots brutos nao sao retornados.
      operationId: getAdminAnuncioHistoricoModeracao
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: historico de moderacao sanitizado
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminModeracaoHistoricoItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/premium/anuncios/{id}:
    get:
      summary: Status Premium read-only de anuncio
      description: Retorna calculo local sanitizado de Premium/beneficios para ADMIN ou MODERADOR. Nao compra, nao ativa, nao cobra, nao chama Pix/Efi, nao retorna valor pago, saldo, credito real, payload de pagamento, telefone bruto, storage ou documento privado.
      operationId: getAdminPremiumAnuncioStatus
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: status Premium sanitizado e somente leitura
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPremiumAnuncioStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/desempenho/anuncios/{id}:
    get:
      summary: Desempenho read-only de anuncio
      description: Retorna prova de resultado local e sanitizada para um anuncio, com visualizacoes, cliques WhatsApp permitidos, taxa clique/view, origens agregadas e comparativo organico/Premium. Nao expoe IP, User-Agent, hash interno, referer bruto, telefone/WhatsApp bruto, storage, pagamento ou credito. Nao executa pixel, tracking externo, cobranca, impulsionamento ou mutacao.
      operationId: getAdminDesempenhoAnuncio
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: desempenho sanitizado do anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDesempenhoAnuncio"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    PublicSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao publica emitida pelo login; cookie HttpOnly, Secure em HML e SameSite Lax.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao administrativa local emitida por `POST /api/admin/auth/login`; cliente deve usar credentials include e o corpo da resposta nunca retorna token ou cookie.
  headers:
    RequestId:
      description: Identificador opaco de correlação da requisição.
      schema:
        type: string
        example: 11111111-1111-4111-8111-111111111111
  parameters:
...
    CurrentUserAdSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]{1,120}$"
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Identificador opaco enviado pelo cliente para correlação.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]{8,128}$"
    Page:
      name: page
      in: query
      required: false
      description: Página baseada em zero.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Quantidade de itens por página, limitada por endpoint.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Parametro invalido ou rota publica proibida.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    NotFound:
      description: Recurso publico nao encontrado.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Autenticacao necessaria ou credenciais invalidas.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Usuario autenticado sem permissao suficiente.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Conflict:
      description: Estado atual impede a transicao solicitada.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    PayloadTooLarge:
      description: Arquivo acima do limite permitido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    UnsupportedMediaType:
      description: Formato binario ou extensao nao permitidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    InternalError:
      description: Falha inesperada sem detalhes internos.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
...
    MeuAnuncio:
      type: object
      additionalProperties: false
      required:
        - id
        - slug
        - titulo
        - descricao
        - categoria
        - preco
        - whatsapp
        - locaisAtendimento
        - servicos
        - status
        - statusModeracao
        - localizacao
        - capa
        - midias
        - atualizadoEm
        - acoesPermitidas
        - visualizacoes
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        titulo:
          type: string
        descricao:
          type: string
        categoria:
          type: string
        preco:
          type: [number, "null"]
          format: double
        whatsapp:
          type: [string, "null"]
          description: Contato do proprio anuncio, disponivel somente neste contrato autenticado.
        locaisAtendimento:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [A_COMBINAR, HOTEL_MOTEL, MEU_LOCAL]
        servicos:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [ANAL, ATRIZ_PORNO, FETICHES, MASSAGEM_TANTRICA, ATIVO, BDSM, JOGOS_DE_INTERPRETACAO, ORAL, ATOR_PORNO, EJACULACAO_CORPORAL, MASSAGEM_EROTICA, PASSIVO, NAMORADAS, TRIO, VIDEOCHAMADA]
        status:
          type: string
          enum: [RASCUNHO, PENDENTE_REVISAO, APROVADO, PUBLICADO, PAUSADO, REJEITADO, REMOVIDO]
        statusModeracao:
          type: string
          enum: [NAO_ENVIADO, PENDENTE, APROVADO, REJEITADO]
        localizacao:
          oneOf:
            - $ref: "#/components/schemas/MeuAnuncioLocalizacao"
            - type: "null"
        capa:
          oneOf:
            - $ref: "#/components/schemas/MeuAnuncioCapa"
            - type: "null"
        midias:
          type: array
          description: Midias vinculadas em ordem, somente para leitura. Conteudo restrito ou nao publicavel nunca recebe URL.
          items:
            $ref: "#/components/schemas/MeuAnuncioMidia"
        atualizadoEm:
          type: [string, "null"]
          format: date-time
        acoesPermitidas:
          $ref: "#/components/schemas/MeuAnuncioAcoesPermitidas"
        visualizacoes:
          $ref: "#/components/schemas/VisualizacoesCanonicas"
...
    MeuAnuncioCicloVida:
      type: object
      additionalProperties: false
      required: [id, slug, status, statusModeracao, atualizadoEm, acoesPermitidas]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        status:
          type: string
          enum: [PUBLICADO, PAUSADO, REMOVIDO]
        statusModeracao:
          type: string
          enum: [NAO_ENVIADO, PENDENTE, APROVADO, REJEITADO]
        atualizadoEm:
          type: [string, "null"]
          format: date-time
        acoesPermitidas:
          $ref: "#/components/schemas/MeuAnuncioAcoesPermitidas"
    MeuAnuncioAtualizacaoRequest:
      type: object
      additionalProperties: false
      required: [titulo, descricao, categoria, preco, uf, cidade, bairro, locaisAtendimento, servicos, whatsapp]
      properties:
        titulo:
          type: string
          minLength: 10
          maxLength: 80
        descricao:
          type: string
          minLength: 20
          maxLength: 600
        categoria:
          type: string
          pattern: "^[A-Z0-9_]{3,40}$"
        preco:
          type: [number, "null"]
          minimum: 0.01
          maximum: 999999.99
        uf:
          type: string
          pattern: "^[A-Z]{2}$"
        cidade:
          type: string
          minLength: 2
          maxLength: 80
        bairro:
          type: [string, "null"]
          minLength: 2
          maxLength: 80
        locaisAtendimento:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [A_COMBINAR, HOTEL_MOTEL, MEU_LOCAL]
        servicos:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [ANAL, ATRIZ_PORNO, FETICHES, MASSAGEM_TANTRICA, ATIVO, BDSM, JOGOS_DE_INTERPRETACAO, ORAL, ATOR_PORNO, EJACULACAO_CORPORAL, MASSAGEM_EROTICA, PASSIVO, NAMORADAS, TRIO, VIDEOCHAMADA]
        whatsapp:
          type: [string, "null"]
          pattern: "^\\+[1-9][0-9]{7,14}$"
    MeuAnuncioLocalizacao:
      type: object
      additionalProperties: false
      required: [uf, cidade, cidadeSlug, bairro, bairroSlug]
      properties:
        uf:
          type: [string, "null"]
        cidade:
          type: [string, "null"]
        cidadeSlug:
          type: [string, "null"]
        bairro:
          type: [string, "null"]
        bairroSlug:
          type: [string, "null"]
...
    MinhasMidiasResponse:
      type: object
      additionalProperties: false
      required: [midias, limites]
      properties:
        midias:
          type: array
          items:
            $ref: "#/components/schemas/MinhaMidiaGestao"
        limites:
          $ref: "#/components/schemas/MinhasMidiasLimites"
...
    AdminAnuncioDetalhe:
      allOf:
        - $ref: "#/components/schemas/AdminAnuncioListaItem"
        - type: object
          additionalProperties: false
          properties:
            descricaoResumo:
              type:
                - string
                - "null"
              description: Texto sanitizado e mascarado; pode ser nulo quando ausente na origem.
            descricao:
              type:
                - string
                - "null"
              description: Conteudo comercial integral para moderacao; pode ser nulo quando ausente.
            categoria:
              type:
                - string
                - "null"
            ultimaPublicacaoEm:
              type:
                - string
                - "null"
              format: date-time
            precoInformado:
              type: boolean
            preco:
              type:
                - number
                - "null"
              format: double
              description: Preco informado no anuncio; pode ser nulo quando nao informado.
            whatsapp:
              type:
                - string
                - "null"
              description: Contato do proprio anuncio para analise; pode ser nulo quando nao informado.
            locaisAtendimento:
              type: array
              items:
                type: string
            servicos:
              type: array
              items:
                type: string
...
    AdminPremiumAnuncioStatus:
      type: object
      additionalProperties: false
      properties:
        anuncioId:
          type: string
          format: uuid
        slug:
          type:
            - string
            - "null"
        titulo:
          type:
            - string
            - "null"
          description: Titulo sanitizado.
        premiumAtivo:
          type: boolean
        destaqueAtivo:
          type: boolean
        topoAtivo:
          type: boolean
        possuiStories:
          type: boolean
        possuiMidiaExtra:
          type: boolean
        beneficiosAtivos:
          type: integer
        beneficiosExpirados:
          type: integer
        beneficiosVencendo:
          type: integer
        inconsistenciasTotal:
          type: integer
        codigosConsistencia:
          type: array
          items:
            type: string
        calculadoEm:
          type:
            - string
            - "null"
          format: date-time
        somenteLeitura:
          type: boolean
          const: true
        compraOuAtivacaoRealDisponivel:
          type: boolean
          const: false
        acoesFinanceirasDisponiveis:
          type: boolean
          const: false
        gratuitoLimitadoPorContato:
          type: boolean
          const: false
        valorPago:
          not: {}
          description: Campo proibido.
        saldoCredito:
          not: {}
          description: Campo proibido.
        payloadPagamento:
          not: {}
          description: Campo proibido.
...
    AdminModeracaoHistoricoItem:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        alvoTipo:
          type: string
          enum: [ANUNCIO, REVISAO_ANUNCIO, ANUNCIO_MIDIA]
        alvoId:
          type: string
          format: uuid
        acao:
          type: string
        decisao:
          type:
            - string
            - "null"
        motivo:
          type:
            - string
            - "null"
        status:
          type:
            - string
            - "null"
        atorId:
          type:
            - string
            - "null"
          format: uuid
        requestId:
          type:
            - string
            - "null"
        resultado:
          type:
            - string
            - "null"
        criadoEm:
          type: string
          format: date-time
...
    AdminPaginaAnuncios:
      type: object
      additionalProperties: false
      properties:
        itens:
          type: array
          items:
            $ref: "#/components/schemas/AdminAnuncioListaItem"
        page:
          type: integer
        size:
          type: integer
          maximum: 50
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
    AdminPaginaMidias:
      type: object
      additionalProperties: false
      properties:
        itens:
          type: array
          items:
            $ref: "#/components/schemas/AdminMidiaListaItem"
        page:
          type: integer
        size:
          type: integer
          maximum: 50
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
...
    AnuncioDetalhePublico:
      type: object
      additionalProperties: false
      required:
        - id
        - slug
        - titulo
        - idadeOculta
        - midias
        - destaque
        - topo
        - midiaExtra
        - story
        - contatoDisponivel
        - comLocal
        - fazAnal
        - locaisAtendimento
        - servicos
        - beneficiosPublicos
        - pendenciaContatoPublico
        - seo
        - visualizacoes
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        titulo:
          type: string
        descricao:
          type:
            - string
            - "null"
        preco:
          type:
            - number
            - "null"
        categoria:
          type:
            - string
            - "null"
        username:
          type:
            - string
            - "null"
          description: Identificador publico da anunciante; nao contem e-mail ou contato privado.
        idade:
          type:
            - integer
            - "null"
          minimum: 18
          description: Idade calculada pelo backend; nula quando ausente ou ocultada por beneficio pago ativo.
        idadeOculta:
          type: boolean
          description: Decisao do backend baseada exclusivamente no beneficio pago OCULTAR_IDADE vigente.
        localizacao:
          $ref: "#/components/schemas/LocalizacaoPublica"
        midias:
          type: array
          items:
            $ref: "#/components/schemas/MidiaPublica"
        destaque:
          type: boolean
        topo:
          type: boolean
        midiaExtra:
          type: boolean
        story:
          type: boolean
        contatoDisponivel:
          type: boolean
        comLocal:
          type: boolean
          description: Verdadeiro somente quando locaisAtendimento contem MEU_LOCAL.
        fazAnal:
          type: boolean
          description: Verdadeiro somente quando servicos contem ANAL.
        locaisAtendimento:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [A_COMBINAR, HOTEL_MOTEL, MEU_LOCAL]
        servicos:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [ANAL, ATRIZ_PORNO, FETICHES, MASSAGEM_TANTRICA, ATIVO, BDSM, JOGOS_DE_INTERPRETACAO, ORAL, ATOR_PORNO, EJACULACAO_CORPORAL, MASSAGEM_EROTICA, PASSIVO, NAMORADAS, TRIO, VIDEOCHAMADA]
        beneficiosPublicos:
          type: array
          description: Rotulos publicos sanitizados de beneficios ativos. Nao expoe valor pago, credito, grupo, campanha ou historico.
          items:
            type: string
        contatoPublico:
          description: Somente o backend pode preencher contato publico. O contato de anuncio publico e ativo nao depende de confirmacao de idade ou visibilidade de midia.
          type:
            - string
            - "null"
        pendenciaContatoPublico:
          description: Marcador local enquanto a politica futura de contato publico nao for aprovada.
          type: string
          const: PENDENTE_POLITICA_EXPOSICAO_WHATSAPP_PUBLICO
        anunciaDesde:
          description: Menor data confiavel de publicacao entre os anuncios da anunciante. Nao deriva da criacao da conta nem da renovacao do anuncio atual.
          type:
            - string
            - "null"
          format: date-time
        publicadoEm:
          type:
            - string
            - "null"
          format: date-time
        seo:
          $ref: "#/components/schemas/SeoRotaPublica"
        visualizacoes:
          $ref: "#/components/schemas/VisualizacoesCanonicas"
...